requirements.txt
fastapi==0.99.0
uvicorn==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.12
sqlalchemy==2.1.0
databases==0.5.5
sqlite==3.36.0
pytest==8.3.3
httpx==0.24.1
